Use a table to represent control modify.alters.adds
We were previously just using "Name: { name }, Value: { value }" and
this basically keeps that structure as a table. In order to not need
recursive tables or recursive text, we flatten the parts into a list
which may not quite represent things correctly so I want to play with
this a little bit more.